Clean up response creation
