+++ fog/view/enclave/impl/src/lib.rs
@@ -294,23 +294,6 @@ where
         client_query_request: QueryRequest,
         shard_query_responses: Vec<MultiViewStoreQueryResponse>,
     ) -> Result<QueryResponse> {
-        // Choose any shard query response and use it to supply the skeleton values for
-        // the QueryResponse. The tx_out_search_results and
-        // highest_processed_block_count fields will be set based on all of the
-        // shard query responses.
-        let shard_query_response = shard_query_responses
-            .first()
-            .expect("Shard query responses must have at least one response.");
-        let shard_query_response_plaintext = self.ake.backend_decrypt(
-            &shard_query_response.store_responder_id,
-            &shard_query_response.encrypted_query_response,
-        )?;
-        let mut shard_query_response: QueryResponse =
-            mc_util_serial::decode(&shard_query_response_plaintext).map_err(|e| {
-                log::error!(self.logger, "Could not decode shard query response: {}", e);
-                Error::ProstDecode
-            })?;
-
         let shard_query_responses = shard_query_responses
             .into_iter()
             .map(|multi_view_store_query_response| {
@@ -327,34 +310,41 @@ where
             })
             .collect::<Result<Vec<_>>>()?;
 
-        shard_query_response.tx_out_search_results =
-            Self::get_collated_tx_out_search_results(client_query_request, &shard_query_responses)?;
-        let last_known_data = get_last_known_data(&shard_query_responses);
-        shard_query_response.last_known_block_count = last_known_data.last_known_block_count;
-        shard_query_response.last_known_block_cumulative_txo_count =
-            last_known_data.last_known_block_cumulative_txo_count;
-        let shared_data: CommonShardData = shard_query_responses.as_slice().into();
-        shard_query_response.missed_block_ranges = shared_data.missed_block_ranges;
-        shard_query_response.rng_records = shared_data.rng_records;
-        shard_query_response.decommissioned_ingest_invocations =
-            shared_data.decommissioned_ingest_invocations;
-        shard_query_response.next_start_from_user_event_id =
-            shared_data.next_start_from_user_event_id;
+        Self::create_response(client_query_request, shard_query_responses)
+    }
 
-        let block_data = get_block_data(shard_query_responses);
-        shard_query_response.highest_processed_block_count =
-            block_data.highest_processed_block_count;
-        shard_query_response.highest_processed_block_signature_timestamp =
+    fn create_response(
+        client_query_request: QueryRequest,
+        mut responses: Vec<DecryptedMultiViewStoreQueryResponse>,
+    ) -> Result<QueryResponse> {
+        let mut result: QueryResponse = QueryResponse::default();
+        let last_known_data = LastKnownData::from(responses.as_slice());
+        result.last_known_block_count = last_known_data.last_known_block_count;
+        result.last_known_block_cumulative_txo_count =
+            last_known_data.last_known_block_cumulative_txo_count;
+
+        let block_data = BlockData::from(responses.as_mut_slice());
+        result.highest_processed_block_count = block_data.highest_processed_block_count;
+        result.highest_processed_block_signature_timestamp =
             block_data.highest_processed_block_signature_timestamp;
 
-        Ok(shard_query_response)
+        let shared_data: CommonShardData = CommonShardData::from(responses.as_slice());
+        result.missed_block_ranges = shared_data.missed_block_ranges;
+        result.rng_records = shared_data.rng_records;
+        result.decommissioned_ingest_invocations = shared_data.decommissioned_ingest_invocations;
+        result.next_start_from_user_event_id = shared_data.next_start_from_user_event_id;
+
+        result.tx_out_search_results =
+            Self::get_collated_tx_out_search_results(client_query_request, &responses)?;
+
+        Ok(result)
     }
 
     fn get_collated_tx_out_search_results(
         client_query_request: QueryRequest,
-        shard_query_responses: &[DecryptedMultiViewStoreQueryResponse],
+        responses: &[DecryptedMultiViewStoreQueryResponse],
     ) -> Result<Vec<TxOutSearchResult>> {
-        let plaintext_search_results = shard_query_responses
+        let plaintext_search_results = responses
             .iter()
             .flat_map(|response| response.query_response.tx_out_search_results.clone())
             .collect::<Vec<TxOutSearchResult>>();
@@ -363,559 +353,5 @@ where
             client_query_request.get_txos,
             plaintext_search_results,
         )
-    }
-}
-
-fn get_block_data(mut responses: Vec<DecryptedMultiViewStoreQueryResponse>) -> BlockData {
-    responses.sort_unstable_by_key(|response| response.block_range.start_block);
-
-    // Find the first time in which a highest processed block count does not equate
-    // to the final block that the shard is responsible for.
-    let mut result = BlockData::default();
-    for response in responses.iter() {
-        let response_highest_processed_block_count =
-            response.query_response.highest_processed_block_count;
-        if response_highest_processed_block_count > result.highest_processed_block_count {
-            result = BlockData::new(
-                response_highest_processed_block_count,
-                response
-                    .query_response
-                    .highest_processed_block_signature_timestamp,
-            );
-        }
-
-        // In this case, the shard hasn't processed all the blocks it's responsible for,
-        // and, as such, those blocks might not be processed so we should return this
-        // number.
-        // TODO: Consider implementing logic that accounts for overlapping block ranges.
-        //   If ranges overlap, then the next server might have processed those blocks
-        //   that this shard did not process (but is responsible for).
-        if response_highest_processed_block_count < response.block_range.end_block {
-            return result;
-        }
-    }
-
-    result
-}
-
-fn get_last_known_data(responses: &[DecryptedMultiViewStoreQueryResponse]) -> LastKnownData {
-    responses
-        .iter()
-        .max_by_key(|response| response.query_response.last_known_block_count)
-        .map_or_else(LastKnownData::default, |response| {
-            LastKnownData::new(
-                response.query_response.last_known_block_count,
-                response
-                    .query_response
-                    .last_known_block_cumulative_txo_count,
-            )
-        })
-}
-
-#[cfg(test)]
-mod get_block_data_tests {
-    use crate::{get_block_data, DecryptedMultiViewStoreQueryResponse};
-    use alloc::{vec, vec::Vec};
-    use mc_fog_types::{common::BlockRange, view::QueryResponse};
-
-    fn create_query_response(
-        highest_processed_block_count: u64,
-        highest_processed_block_signature_timestamp: u64,
-    ) -> QueryResponse {
-        QueryResponse {
-            highest_processed_block_count,
-            highest_processed_block_signature_timestamp,
-            next_start_from_user_event_id: 0,
-            missed_block_ranges: vec![],
-            rng_records: vec![],
-            decommissioned_ingest_invocations: vec![],
-            tx_out_search_results: vec![],
-            last_known_block_count: highest_processed_block_count,
-            last_known_block_cumulative_txo_count: 0,
-        }
-    }
-
-    #[test]
-    fn all_responses_fully_processed() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-        for i in 0..STORE_COUNT {
-            let query_response = create_query_response((i + 1) as u64, i as u64);
-            let block_range = BlockRange::new(i as u64, (i + 1) as u64);
-            let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-                query_response,
-                block_range,
-            };
-            decrypted_query_responses.push(decrypted_query_response);
-        }
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        let last_response = decrypted_query_responses.last().unwrap();
-        assert_eq!(
-            result.highest_processed_block_count,
-            last_response.query_response.highest_processed_block_count
-        );
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            last_response
-                .query_response
-                .highest_processed_block_signature_timestamp
-        );
-    }
-
-    #[test]
-    fn first_response_incomplete() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response "incomplete"- i.e. it hasn't processed all of its
-        // blocks.
-        let incomplete_query_response = create_query_response(2, 2);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response.clone(),
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response fully processed.
-        let query_response = create_query_response(6, 6);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the third response fully processed.
-        let query_response = create_query_response(9, 9);
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response fully processed.
-        let query_response = create_query_response(12, 12);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(
-            result.highest_processed_block_count,
-            incomplete_query_response.highest_processed_block_count
-        );
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            incomplete_query_response.highest_processed_block_signature_timestamp
-        );
-    }
-
-    #[test]
-    fn second_response_zero_processed_blocks() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let fully_processed_block_count = 3;
-        let fully_processed_timestamp = 3;
-        let query_response =
-            create_query_response(fully_processed_block_count, fully_processed_timestamp);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response process zero blocks.
-        let query_response = create_query_response(0, 0);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the third response fully processed.
-        let query_response = create_query_response(9, 9);
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response incomplete.
-        let query_response = create_query_response(10, 10);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(
-            result.highest_processed_block_count,
-            fully_processed_block_count
-        );
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            fully_processed_timestamp
-        );
-    }
-    #[test]
-    fn second_response_incomplete() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let query_response = create_query_response(3, 3);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response "incomplete"- i.e. it hasn't processed all of its
-        // blocks.
-        let incomplete_block_count = 4;
-        let incomplete_timestamp = 4;
-        let incomplete_query_response =
-            create_query_response(incomplete_block_count, incomplete_block_count);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response,
-            block_range,
-        });
-
-        // Make the third response fully processed.
-        let query_response = create_query_response(9, 9);
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response incomplete.
-        let query_response = create_query_response(10, 10);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(result.highest_processed_block_count, incomplete_block_count);
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            incomplete_timestamp
-        );
-    }
-
-    #[test]
-    fn penultimate_response_incomplete() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let query_response = create_query_response(3, 3);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response fully processed.
-        let incomplete_query_response = create_query_response(6, 6);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response,
-            block_range,
-        });
-
-        // Make the third response incomplete.
-        let incomplete_block_count = 8;
-        let incomplete_timestamp = 8;
-        let query_response = create_query_response(incomplete_block_count, incomplete_timestamp);
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response fully processed.
-        let query_response = create_query_response(12, 12);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(result.highest_processed_block_count, incomplete_block_count);
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            incomplete_timestamp
-        );
-    }
-
-    #[test]
-    fn penultimate_response_zero_processed_blocks() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let query_response = create_query_response(1, 1);
-        let block_range = BlockRange::new(0, 1);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response fully processed.
-        let second_response_highest_processed_block_count = 2;
-        let second_response_timestamp = 2;
-        let query_response = create_query_response(
-            second_response_highest_processed_block_count,
-            second_response_timestamp,
-        );
-        let block_range = BlockRange::new(1, 2);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the third response process zero blocks.
-        let incomplete_query_response = create_query_response(0, 0);
-        let block_range = BlockRange::new(2, 3);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response,
-            block_range,
-        });
-
-        // Make the fourth response fully processed.
-        let query_response = create_query_response(4, 4);
-        let block_range = BlockRange::new(3, 4);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(
-            result.highest_processed_block_count,
-            second_response_highest_processed_block_count,
-        );
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            second_response_timestamp
-        );
-    }
-
-    #[test]
-    fn final_response_incomplete() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let query_response = create_query_response(3, 3);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response fully processed.
-        let incomplete_query_response = create_query_response(6, 6);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response,
-            block_range,
-        });
-
-        // Make the third response fully processed.
-        let query_response = create_query_response(9, 9);
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response incomplete.
-        let incomplete_block_count = 10;
-        let incomplete_timestamp = 10;
-        let query_response = create_query_response(10, 10);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(result.highest_processed_block_count, incomplete_block_count);
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            incomplete_timestamp
-        );
-    }
-
-    #[test]
-    fn final_response_zero_processed_blocks() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        // Make the first response fully processed.
-        let query_response = create_query_response(3, 3);
-        let block_range = BlockRange::new(0, 3);
-        let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        };
-        decrypted_query_responses.push(decrypted_query_response);
-
-        // Make the second response fully processed.
-        let incomplete_query_response = create_query_response(6, 6);
-        let block_range = BlockRange::new(3, 6);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response: incomplete_query_response,
-            block_range,
-        });
-
-        // Make the third response fully processed.
-        let last_fully_processed_block_count = 9;
-        let last_fully_processed_timestamp = 9;
-        let query_response = create_query_response(
-            last_fully_processed_block_count,
-            last_fully_processed_timestamp,
-        );
-        let block_range = BlockRange::new(6, 9);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        // Make the fourth response incomplete.
-        let query_response = create_query_response(0, 0);
-        let block_range = BlockRange::new(9, 12);
-        decrypted_query_responses.push(DecryptedMultiViewStoreQueryResponse {
-            query_response,
-            block_range,
-        });
-
-        let result = get_block_data(decrypted_query_responses.clone());
-
-        assert_eq!(
-            result.highest_processed_block_count,
-            last_fully_processed_block_count
-        );
-        assert_eq!(
-            result.highest_processed_block_signature_timestamp,
-            last_fully_processed_timestamp
-        );
-    }
-}
-#[cfg(test)]
-mod last_known_block_data_tests {
-    use crate::{get_last_known_data, DecryptedMultiViewStoreQueryResponse};
-    use alloc::{vec, vec::Vec};
-    use mc_fog_types::{common::BlockRange, view::QueryResponse};
-
-    fn create_query_response(
-        last_known_block_count: u64,
-        last_known_block_cumulative_txo_count: u64,
-    ) -> QueryResponse {
-        QueryResponse {
-            highest_processed_block_count: 0,
-            highest_processed_block_signature_timestamp: 0,
-            next_start_from_user_event_id: 0,
-            missed_block_ranges: vec![],
-            rng_records: vec![],
-            decommissioned_ingest_invocations: vec![],
-            tx_out_search_results: vec![],
-            last_known_block_count,
-            last_known_block_cumulative_txo_count,
-        }
-    }
-
-    #[test]
-    fn different_last_known_block_counts() {
-        const STORE_COUNT: usize = 4;
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-
-        for i in 0..STORE_COUNT {
-            let last_known_block_count = ((i + 1) * 10) as u64;
-            let last_known_block_cumulative_txo_count = last_known_block_count * 2;
-            let query_response = create_query_response(
-                last_known_block_count,
-                last_known_block_cumulative_txo_count,
-            );
-            let block_range = BlockRange::new(i as u64, last_known_block_count);
-            let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-                query_response,
-                block_range,
-            };
-            decrypted_query_responses.push(decrypted_query_response);
-        }
-
-        let last_response = decrypted_query_responses
-            .last()
-            .expect("Couldn't get last decrypted query response");
-        let expected_last_known_block_count = last_response.query_response.last_known_block_count;
-        let expected_last_known_block_cumulative_txo_count = last_response
-            .query_response
-            .last_known_block_cumulative_txo_count;
-
-        let result = get_last_known_data(&decrypted_query_responses);
-
-        assert_eq!(
-            result.last_known_block_count,
-            expected_last_known_block_count
-        );
-        assert_eq!(
-            result.last_known_block_cumulative_txo_count,
-            expected_last_known_block_cumulative_txo_count
-        );
-    }
-
-    #[test]
-    fn same_last_known_block_counts() {
-        const STORE_COUNT: usize = 4;
-        const LAST_KNOWN_BLOCK_COUNT: u64 = 100;
-        const LAST_KNOWN_BLOCK_CUMULATIVE_TXO_COUNT: u64 = 1000;
-
-        let mut decrypted_query_responses = Vec::with_capacity(STORE_COUNT);
-        for i in 0..STORE_COUNT {
-            let end_block_count = ((i + 1) * 25) as u64;
-            let query_response = create_query_response(
-                LAST_KNOWN_BLOCK_COUNT,
-                LAST_KNOWN_BLOCK_CUMULATIVE_TXO_COUNT,
-            );
-            let block_range = BlockRange::new(i as u64, end_block_count);
-            let decrypted_query_response = DecryptedMultiViewStoreQueryResponse {
-                query_response,
-                block_range,
-            };
-            decrypted_query_responses.push(decrypted_query_response);
-        }
-
-        let result = get_last_known_data(&decrypted_query_responses);
-
-        assert_eq!(result.last_known_block_count, LAST_KNOWN_BLOCK_COUNT);
-        assert_eq!(
-            result.last_known_block_cumulative_txo_count,
-            LAST_KNOWN_BLOCK_CUMULATIVE_TXO_COUNT
-        );
     }
 }
